ocr-table-extract-main/uniocr_ai/requirements.txt
cycler==0.12.1
easydict==1.9
editdistance==0.8.1
Flask==1.1.2
lmdb==1.3.0
itsdangerous==2.0.1
Jinja2==2.11.3
jpype1==1.5.0
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.24.4
opencv-python==4.5.5.64
pdfrw==0.4
Pillow==9.1.0
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyJWT==2.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reportlab==3.6.11
requests==2.26.0
scipy==1.10.1
six==1.16.0
scikit-image==0.19.3
urllib3==1.26.20